open weather data downloading, rest api tests

+++ src/main/java/cz/moravec/repository/TownRepository.java
@@ -1,13 +1,9 @@
 package cz.moravec.repository;
 
 import cz.moravec.model.Town;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import java.util.List;
 
 @Repository
-public interface TownRepository extends JpaRepository<Town,Integer> {
+public interface TownRepository extends JpaRepository<Town,Long> {
 }
